Set emoji keyboard height to the same size as the softkeyboard. Store the height for later use
