+++ app/src/main/AndroidManifest.xml
@@ -45,6 +45,7 @@
 
         <activity
             android:name=".chatroom.ui.ChatRoomActivity"
+            android:windowSoftInputMode="adjustPan"
             android:theme="@style/AppTheme" />
 
         <activity
